adding the class support
